refactor: simplify deployment skip patterns
